Bris branch merge to Zorro

+++ AllNotes.Domain/Models/Sport/Series.cs
@@ -8,7 +8,7 @@ namespace AllNotes.Domain.Models.Sport
     public class Series : BaseEntity
     {
         public int Repeats { get; set; }
-        public int Weights { get; set; }
+        public float Weights { get; set; }
         public int ExerciseId { get; set; }
         public virtual Exercise Exercise { get; set; }
     }
